# backend/app/api/v1/alerts.py
"""
Alert management endpoints.

List, inspect, and acknowledge security alerts generated by the
detection pipeline.
"""

from __future__ import annotations

import logging

from fastapi import APIRouter, Depends, HTTPException, Query, status

from app.dependencies import CurrentUser, DBSession
from app.schemas.alert import AlertAcknowledgeRequest, AlertListResponse, AlertResponse
from app.services.alert_service import AlertService

logger = logging.getLogger("agentshield.api.alerts")

router = APIRouter()


# --------------------------------------------------------------------------
# GET /
# --------------------------------------------------------------------------


@router.get(
    "/",
    response_model=AlertListResponse,
    status_code=status.HTTP_200_OK,
    summary="List alerts",
    description=(
        "Return a filtered, paginated list of security alerts. "
        "Supports filtering by severity and acknowledgement status."
    ),
)
async def list_alerts(
    current_user: CurrentUser,
    db: DBSession,
    severity: str | None = Query(
        default=None,
        description="Filter by severity (critical/high/medium/low)",
    ),
    acknowledged: bool | None = Query(
        default=None,
        description="Filter by acknowledgement status",
    ),
    alert_type: str | None = Query(
        default=None,
        description="Filter by alert type",
    ),
    session_id: str | None = Query(default=None, description="Filter by session ID"),
    agent_id: str | None = Query(default=None, description="Filter by agent ID"),
    page: int = Query(default=1, ge=1, description="Page number"),
    size: int = Query(default=20, ge=1, le=100, description="Page size"),
) -> AlertListResponse:
    """List security alerts with optional filters."""
    logger.info(
        "GET /alerts  user=%s  severity=%s  ack=%s  page=%d",
        current_user["user_id"],
        severity,
        acknowledged,
        page,
    )

    filters: dict[str, object] = {}
    if severity:
        filters["severity"] = severity
    if acknowledged is not None:
        filters["is_acknowledged"] = acknowledged
    if alert_type:
        filters["alert_type"] = alert_type
    if session_id:
        filters["session_id"] = session_id
    if agent_id:
        filters["agent_id"] = agent_id

    service = AlertService(db_session=db)

    try:
        result = await service.get_alerts(filters=filters, page=page, size=size)
    except Exception as exc:
        logger.exception("Failed to list alerts")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to list alerts: {exc}",
        ) from exc

    return result


# --------------------------------------------------------------------------
# GET /{alert_id}
# --------------------------------------------------------------------------


@router.get(
    "/{alert_id}",
    response_model=AlertResponse,
    status_code=status.HTTP_200_OK,
    summary="Get alert",
    description="Retrieve a single security alert by its ID.",
)
async def get_alert(
    alert_id: str,
    current_user: CurrentUser,
    db: DBSession,
) -> AlertResponse:
    """Fetch a single alert by ID."""
    logger.info(
        "GET /alerts/%s  user=%s",
        alert_id,
        current_user["user_id"],
    )

    service = AlertService(db_session=db)

    try:
        result = await service.get_alert(alert_id)
    except ValueError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc
    except Exception as exc:
        logger.exception("Failed to get alert %s", alert_id)
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to retrieve alert: {exc}",
        ) from exc

    return result


# --------------------------------------------------------------------------
# POST /{alert_id}/acknowledge
# --------------------------------------------------------------------------


@router.post(
    "/{alert_id}/acknowledge",
    response_model=AlertResponse,
    status_code=status.HTTP_200_OK,
    summary="Acknowledge alert",
    description=(
        "Mark a security alert as acknowledged. Records who acknowledged "
        "it and the timestamp."
    ),
)
async def acknowledge_alert(
    alert_id: str,
    body: AlertAcknowledgeRequest,
    current_user: CurrentUser,
    db: DBSession,
) -> AlertResponse:
    """Acknowledge a security alert."""
    acknowledged_by = body.acknowledged_by or str(current_user["user_id"])

    logger.info(
        "POST /alerts/%s/acknowledge  user=%s  by=%s",
        alert_id,
        current_user["user_id"],
        acknowledged_by,
    )

    service = AlertService(db_session=db)

    try:
        result = await service.acknowledge_alert(
            alert_id=alert_id,
            acknowledged_by=acknowledged_by,
        )
    except ValueError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc
    except Exception as exc:
        logger.exception("Failed to acknowledge alert %s", alert_id)
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to acknowledge alert: {exc}",
        ) from exc

    return result